Move BlockManager to node/blockstorage
Can be reviewed with --color-moved=dimmed-zebra
